Create robottelo.utils, move issue_handler and version helpers
Create pytest_plugin.issue_handlers module
Fix behavior of bz-cache option, made it a flag option, always using default file
Create version util
Move things from helpers into specific modules

Move pytest hook for issue handler, move is_open to util

Added basic unit test of file creation behavior

Handled slugify exception when processing docblocks for unit tests
