+++ tests/foreman/api/test_reporttemplates.py
@@ -34,8 +34,8 @@ from robottelo.decorators import stubbed
 from robottelo.decorators import tier1
 from robottelo.decorators import tier2
 from robottelo.decorators import tier3
-from robottelo.helpers import is_open
 from robottelo.test import APITestCase
+from robottelo.utils.issue_handlers import is_open
 from robottelo.vm import VirtualMachine
