+++ tests/foreman/api/test_syncplan.py
@@ -47,8 +47,8 @@ from robottelo.decorators import tier2
 from robottelo.decorators import tier3
 from robottelo.decorators import tier4
 from robottelo.decorators import upgrade
-from robottelo.helpers import is_open
 from robottelo.test import APITestCase
+from robottelo.utils.issue_handlers import is_open
 
 
 @filtered_datapoint
