+++ tests/foreman/cli/test_activationkey.py
@@ -56,9 +56,9 @@ from robottelo.decorators import tier1
 from robottelo.decorators import tier2
 from robottelo.decorators import tier3
 from robottelo.decorators import upgrade
-from robottelo.helpers import is_open
 from robottelo.ssh import upload_file
 from robottelo.test import CLITestCase
+from robottelo.utils.issue_handlers import is_open
 from robottelo.vm import VirtualMachine
